refactor: restructure components for better organization
