+++ Frontend/src/components/common/Navbar.jsx
@@ -5,8 +5,8 @@ import { IoIosSearch, IoIosHeartEmpty } from "react-icons/io";
 import { IoBagHandleOutline } from "react-icons/io5";
 import { Link, useLocation } from "react-router-dom";
 import { useTranslation } from "react-i18next";
-import LanguageToggleSwitch from "../language_toggle/switch_toggle/LanguageToggleSwitch";
-import LanguageToggleClick from "../language_toggle/click_toggle/LanguageToggleClick";
+import LanguageToggleSwitch from "./language_toggle/switch_toggle/LanguageToggleSwitch";
+import LanguageToggleClick from "./language_toggle/click_toggle/LanguageToggleClick";
 import { GiHamburgerMenu } from "react-icons/gi";
 import { useNavBehavior } from "../../hook/useNavBehavior";
 import { useLanguageDirection } from "../../hook/useLanguageDirection";
