+++ Frontend/src/components/common/card/product_card/ProductCard.css
