+++ Frontend/src/components/common/language_toggle/switch_toggle/LanguageToggleSwitch.css
