+++ Frontend/src/components/common/language_toggle/switch_toggle/LanguageToggleSwitch.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
-import useIsArabic from "../../../hook/useIsArabic";
-import { useLanguageDirection } from "../../../hook/useLanguageDirection";
+import useIsArabic from "../../../../hook/useIsArabic";
+import { useLanguageDirection } from "../../../../hook/useLanguageDirection";
 import "./LanguageToggleSwitch.css";
 
 function LanguageToggleSwitch(){
