+++ Frontend/src/components/common/reveal_animation/ScrollReveal.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import useScrollReveal from "../../hook/useScrollReveal";
+import useScrollReveal from "../../../hook/useScrollReveal";
 
 interface Props {
   type?: "up" | "down";
