+++ Frontend/src/components/common/star_rating/StarRating.css
