+++ Frontend/src/components/home/banner/Banner.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import BannerImg from "../../../assets/image/home/banner/shopping.avif";
 import { bannerData } from "../../../assets/data/bannerData";
 import useIsArabic from "../../../hook/useIsArabic";
-import ScrollReveal from "../../reveal_animation/ScrollReveal";
-import RevealGroup from "../../reveal_animation/RevealGroup";
+import ScrollReveal from "../../common/reveal_animation/ScrollReveal";
+import RevealGroup from "../../common/reveal_animation/RevealGroup";
 import "./Banner.css";
