+++ Frontend/src/style/components/Footer.css
@@ -40,7 +40,7 @@
 
 .footer-title {
   font-size: 1.2rem;
-  font-weight: 600;
+  font-weight: 60;
   margin-bottom: 15px;
   color: #fff;
 }
